+++ src/components/Card/index.jsx
@@ -1,4 +1,4 @@
-import styles from './Card.module.css'
+import styles from './styles.module.css'
 
 const Card = ({ info }) => {
     return (
