+++ src/components/ui/Button/index.jsx
@@ -1,4 +1,4 @@
-import styles from './Button.module.css'
+import styles from './styles.module.css'
 
 const Button = ({ onClick, value }) => {
     return (
